Fix for nav error in question pool.
git-svn-id: http://svn.ilias.de/svn/ilias/branches/Release_4_3_x_branch@54146 21b2c9ec-7c21-0410-8b45-9bfb8ed2bfc5

+++ Modules/TestQuestionPool/classes/class.ilObjQuestionPoolGUI.php
@@ -892,8 +892,10 @@ class ilObjQuestionPoolGUI extends ilObjectGUI
 
 			}
 		}
-
-		$this->object->purgeQuestions();
+		if($this->object instanceof ilObQuestionPool)
+		{
+			$this->object->purgeQuestions();
+		}
 		// reset test_id SESSION variable
 		$_SESSION["test_id"] = "";
